requirements.txt
requests
unstructured[html]
chromadb
ollama
python-dotenv
